move removeCircularRefs to workflow/src/utils
