+++ packages/workflow/src/utils.ts
@@ -1,4 +1,4 @@
-import type { BinaryFileType } from './Interfaces';
+import type { BinaryFileType, JsonObject } from './Interfaces';
 
 const readStreamClasses = new Set(['ReadStream', 'Readable', 'ReadableStream']);
 
@@ -127,3 +127,35 @@ export function assert<T>(condition: T, msg?: string): asserts condition {
 		throw error;
 	}
 }
+
+export const isTraversableObject = (value: any): value is JsonObject => {
+	return (
+		value &&
+		typeof value === 'object' &&
+		!Array.isArray(value) &&
+		// eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+		!!Object.keys(value).length
+	);
+};
+
+export const removeCircularRefs = (obj: JsonObject, seen = new Set()) => {
+	seen.add(obj);
+	Object.entries(obj).forEach(([key, value]) => {
+		if (isTraversableObject(value)) {
+			// eslint-disable-next-line @typescript-eslint/no-unused-expressions
+			seen.has(value) ? (obj[key] = { circularReference: true }) : removeCircularRefs(value, seen);
+			return;
+		}
+		if (Array.isArray(value)) {
+			value.forEach((val, index) => {
+				if (seen.has(val)) {
+					value[index] = { circularReference: true };
+					return;
+				}
+				if (isTraversableObject(val)) {
+					removeCircularRefs(val, seen);
+				}
+			});
+		}
+	});
+};
